+++ modules/sheets/tagmarAltSheet.js
@@ -3,10 +3,10 @@ export default class tagmarAltSheet extends ActorSheet {
     static get defaultOptions() {
         this.lastUpdate = {};
         this.lastItemsUpdate = [];
-        return mergeObject(super.defaultOptions, {
+        return foundry.utils.mergeObject(super.defaultOptions, {
         classes: ["tagmar", "sheet", "actor"],
         //width: 900,
-        height: 855,
+        height: 890,
         tabs: [{
             navSelector: ".prim-tabs",
             contentSelector: ".sheet-primary",
@@ -324,7 +324,22 @@ export default class tagmarAltSheet extends ActorSheet {
             html.find('.searchHabilidade').keyup(this._realcaHablidade.bind(this));
             html.find('.searchEfeito').keyup(this._realcaEfeito.bind(this));
             html.find('.descansar').click(this._descanso.bind(this));
+            html.find('input[name="grupo_aprim"]').click(this._radio_grupo_aprim.bind(this));
         } 
+    }
+
+    _radio_grupo_aprim(event) {
+        event.preventDefault();
+        let old_grup = this.document.system.grupo_aprim;
+        if (old_grup == $(event.currentTarget).val()) {
+            this.document.update({
+                'system.grupo_aprim': ""
+            });
+        } else {
+            this.document.update({
+                'system.grupo_aprim': $(event.currentTarget).val()
+            });
+        }
     }
 
     _linguasDialog(event) {
@@ -1271,7 +1286,7 @@ export default class tagmarAltSheet extends ActorSheet {
         const h_sub = actorData.items.filter(item => item.type == "Habilidade" && item.system.tipo == "subterfugio");
         const h_inf = actorData.items.filter(item => item.type == "Habilidade" && item.system.tipo == "influencia");
         const h_geral = actorData.items.filter(item => item.type == "Habilidade" && item.system.tipo == "geral");
-        const tecnicas = actorData.items.filter(item => item.type == "TecnicasCombate");
+        const tecnicas = actorData.items.filter(item => item.type == "Tecnica_Combate");
         const defesas = actorData.items.filter(item => item.type == "Defesa");
         const transportes = actorData.items.filter(item => item.type == "Transporte");
         const pertences = actorData.items.filter(item => item.type == "Pertence" && !item.system.inTransport);
@@ -1280,6 +1295,14 @@ export default class tagmarAltSheet extends ActorSheet {
         const profissoes = actorData.items.filter(item => item.type == "Profissao");
         //if (racas.length >= 1) this.document.deleteEmbeddedDocuments("Item", [racas]);
         //if (profissoes.length >= 1) this.document.deleteEmbeddedDocuments("Item", [item._id]);
+        const old_tecnicas = actorData.items.filter(item => item.type == "TecnicasCombate");
+        if (old_tecnicas.length > 0) {
+            let old_ids = [];
+            old_tecnicas.forEach(function (old) {
+                old_ids.push(old.id);
+            });
+            this.document.deleteEmbeddedDocuments("Item", old_ids);
+        }
         var especializacoes = [];
         const efeitos = actorData.items.filter(item => item.type == "Efeito");
         const tabela_resol = [
@@ -1338,7 +1361,10 @@ export default class tagmarAltSheet extends ActorSheet {
             [20,  2,  2,  2,  2,  2,  2,  3,  3,  4,  4,  5,  5,  6,  6,  7,  7,  8,  9, 10, 11]
         ];
         if (profissoes[0]) {
-            especializacoes = profissoes[0].system.especializacoes.split(",");
+            //especializacoes = profissoes[0].system.especializacoes.split(",");
+            profissoes[0].system.especializacoes.split(",").forEach(function (espec) {
+                especializacoes.push({key : espec});
+            });
         } // Alow
         
         if (h_prof.length > 1) h_prof.sort(function (a, b) {
@@ -1367,9 +1393,6 @@ export default class tagmarAltSheet extends ActorSheet {
         });
         if (tecnicas.length > 1) tecnicas.sort(function (a, b) {
             return a.name.localeCompare(b.name);
-        });
-        if (tecnicas.length > 1) tecnicas.sort(function (a, b) {
-            return a.system.categoria.localeCompare(b.system.categoria);
         });
         if (defesas.length > 1) defesas.sort(function (a, b) {
             return a.name.localeCompare(b.name);
@@ -1409,12 +1432,98 @@ export default class tagmarAltSheet extends ActorSheet {
         actorData.h_sub = h_sub;
         actorData.h_inf = h_inf;
         actorData.h_geral = h_geral;
+        actorData.habilidades = h_prof.concat(h_man, h_con, h_sub, h_inf, h_geral);
+        actorData.cat_def = [{key : "L"}, {key : "M"}, {key : "P"}];
         actorData.combate = combate;
         actorData.combate_fav = combate_fav;
         actorData.tecnica_fav = tecnica_fav;
         actorData.magia_fav = magia_fav;
         actorData.magias = magias;
         actorData.ficha = "Pontos";
+        actorData.cd_aprim = false;
+        actorData.ci_aprim = false;
+        actorData.cl_aprim = false;
+        actorData.cld_aprim = false;
+        actorData.el_aprim = false;
+        actorData.cme_aprim = false;
+        actorData.cmm_aprim = false;
+        actorData.em_aprim = false;
+        actorData.pma_aprim = false;
+        actorData.pml_aprim = false;
+        actorData.cpe_aprim = false;
+        actorData.cpm_aprim = false;
+        actorData.ep_aprim = false;
+        actorData.pp_aprim = false;
+        actorData.ppa_aprim = false;
+        actorData.ppb_aprim = false;
+        switch (actorData.system.grupo_aprim) {
+            case "CD":
+                actorData.cd_aprim = true;
+                break;
+            case "CI":
+                actorData.ci_aprim = true;
+                break;
+            case "CL":
+                actorData.cl_aprim = true;
+                break;
+            case "CLD":
+                actorData.cld_aprim = true;
+                break;
+            case "EL":
+                actorData.el_aprim = true;
+                break;
+            case "CmE":
+                actorData.cme_aprim = true;
+                break;
+            case "CmM":
+                actorData.cmm_aprim = true;
+                break;
+            case "EM":
+                actorData.em_aprim = true;
+                break;
+            case "PmA":
+                actorData.pma_aprim = true;
+                break;
+            case "PmL":
+                actorData.pml_aprim = true;
+                break;
+            case "CpE":
+                actorData.cpe_aprim = true;
+                break;
+            case "CpM":
+                actorData.cpm_aprim = true;
+                break;
+            case "EP":
+                actorData.ep_aprim = true;
+                break;
+            case "PP":
+                actorData.pp_aprim = true;
+                break;
+            case "PpA":
+                actorData.ppa_aprim = true;
+                break;
+            case "PpB":
+                actorData.ppb_aprim = true;
+                break;
+            default:
+                actorData.cd_aprim = false;
+                actorData.ci_aprim = false;
+                actorData.cl_aprim = false;
+                actorData.cld_aprim = false;
+                actorData.el_aprim = false;
+                actorData.cme_aprim = false;
+                actorData.cmm_aprim = false;
+                actorData.em_aprim = false;
+                actorData.pma_aprim = false;
+                actorData.pml_aprim = false;
+                actorData.cpe_aprim = false;
+                actorData.cpm_aprim = false;
+                actorData.ep_aprim = false;
+                actorData.pp_aprim = false;
+                actorData.ppa_aprim = false;
+                actorData.ppb_aprim = false;
+                break;
+        }
     }
 
 }