+++ modules/tagmar.js
@@ -912,7 +912,7 @@ Hooks.on('tagmar_Critico', async function (coluna, tabela_resol, user, actor, ti
 
 async function rolarCritico(coluna, tabela_resol, user, actor, tipo, falha) {
   let roll = new Roll('1d20');
-  roll.evaluate({async: false});
+  await roll.evaluate();
   let result = roll.total;
   let conteudo = "<h1 class='mediaeval rola' style='text-align:center;'>Rolagem do Crítico</h1><br><p class='mediaeval rola_desc'>";
   let table = "";
@@ -1347,7 +1347,7 @@ async function rollTabela(colunaR) {
   let r = new Roll("1d20");
   let resultado = "";
   let PrintResult = "";
-  r.evaluate({async: false});
+  await r.evaluate();
   var Dresult = r.total;
   let coluna_table = tabela_resol.find(h => h[0] == colunaR);
   resultado = coluna_table[Dresult];
@@ -1403,7 +1403,7 @@ async function rollResistencia(resist, f_ataque) {
     else if (def_ataq <= -16) valorSucess = 20;
   }
   const r = new Roll("1d20");
-  r.evaluate({async: false});
+  await r.evaluate();
   const Dresult = r.total;
   if ((Dresult >= valorSucess || Dresult == 20) && Dresult > 1) { // Sucesso
       stringSucesso = "<h1 class='mediaeval rola' style='text-align:center; color: white;background-color:#00a1e8;'>SUCESSO</h1>";
@@ -1490,6 +1490,9 @@ Hooks.on("renderCombatTracker",function (combatTracker, html) {
 async function createTagmarMacro(data, slot) {
   if (data.type !== "Item") return;
   let data_a = data.uuid.split('.');
+  if (data_a.length != 4) {
+    return ui.notifications.error("Verifique se o token está linkado com a ficha.");
+  }
   let item_id = data_a[data_a.length-1];
   let actor_id = data_a[1];
   const item = game.actors.get(actor_id).items.get(item_id);